+++ ReleaseNotes.md
@@ -4,3 +4,4 @@
 * Adds `Get-NamingConventions` which is a utility function that produces name variations from an upper camel case name.
 * Adds `New-AngularComponent` scaffolding function that creates the associated typescript, html, style, and spec files.
 * Adds `New-AngularFeature` scaffolding function that creates the files for a module, bundle, and optional routing module.
+* Adds `New-AngularService` scaffolding function that creates a basic service and associated spec file.